+++ src/main/index.ts
@@ -33,7 +33,7 @@ export {
   withScope,
 } from '@sentry/core';
 
-import { initAndBind } from '@sentry/core';
+import { getCurrentHub, initAndBind } from '@sentry/core';
 import { _callOnClient } from '@sentry/minimal';
 import { defaultIntegrations } from '@sentry/node';
 import { Event } from '@sentry/types';
@@ -86,4 +86,32 @@ export function showReportDialog(): void {
  */
 export function captureMinidump(path: string, event: Event = {}): void {
   _callOnClient('captureMinidump', path, event);
+}
+
+/**
+ * A promise that resolves when all current events have been sent.
+ * If you provide a timeout and the queue takes longer to drain the promise returns false.
+ *
+ * @param timeout Maximum time in ms the client should wait.
+ */
+export async function flush(timeout?: number): Promise<boolean> {
+  const client = getCurrentHub().getClient<MainClient>();
+  if (client) {
+    return client.flush(timeout);
+  }
+  return Promise.reject(false);
+}
+
+/**
+ * A promise that resolves when all current events have been sent.
+ * If you provide a timeout and the queue takes longer to drain the promise returns false.
+ *
+ * @param timeout Maximum time in ms the client should wait.
+ */
+export async function close(timeout?: number): Promise<boolean> {
+  const client = getCurrentHub().getClient<MainClient>();
+  if (client) {
+    return client.close(timeout);
+  }
+  return Promise.reject(false);
 }